requirements-base.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
pyyaml==6.0.2
numpy==1.26.4
opencv-python==4.10.0.84
requests==2.32.3
httpx==0.28.1
python-dateutil==2.9.0.post0
Pillow==11.0.0
ultralytics==8.3.56
onnxruntime==1.20.1
insightface==0.7.3
torch==2.5.1
torchvision==0.20.1
faiss-cpu==1.9.0.post1 ; platform_system != "Windows"
pytest==8.3.4
